[Test] QueryDsl Join Test

+++ src/main/java/com/sparta/moit/domain/meeting/controller/docs/MeetingControllerDocs.java
@@ -26,8 +26,8 @@ public interface MeetingControllerDocs {
     @Operation(summary = "모임 조회", description = "모임 조회 API")
     ResponseEntity<?> getMeetingListByLatLng(@RequestParam Double locationLat,
                                              @RequestParam Double locationLng,
-                                             @RequestParam(required = false) List<Short> skillId,
-                                             @RequestParam(required = false) List<Short> careerId,
+                                             @RequestParam(required = false) List<Long> skillId,
+                                             @RequestParam(required = false) List<Long> careerId,
                                              @RequestParam(defaultValue = "1") int page);
 
     @Operation(summary = "회원 모임 가입", description = "모임 가입 API")
